\name{ceu1kgv-package}
\alias{ceu1kgv-package}
\alias{ceu1kgv}
\alias{ex}
\docType{package}
\title{
expression + genotype on 79 unrelated CEU individuals
}
\description{
expression + genotype on 79 unrelated CEU individuals

}
\details{
\tabular{ll}{
Package: \tab ceu1kgv\cr
Version: \tab 0.1.0\cr
Depends: \tab GGBase\cr
License: \tab Private\cr
LazyLoad: \tab yes\cr
}

The object eset.rda in data includes an ExpressionSet named 'ex'.
}
\author{
VJ Carey <stvjc@channing.harvard.edu>

Maintainer: VJ Carey <stvjc@channing.harvard.edu>

Expression data from ArrayExpress E-MTAB-264.  
}
\references{
PMID 22532805
}
\keyword{ package }
%\seealso{
%}
%\examples{
%}
